refactor(plugin): PluginManager god-object 迁出 core → app.helper.plugin_manager(S9)
app/core/plugin.py 是 2026 行 god-object,既不该留在 core(高层插件编排不属下层),又顶层反向依赖 db(PluginDataOper/SystemConfigOper)。git mv 迁至 helper 层(helper→{core,db,utils} 均合法层序);core/plugin.py 留 PEP 562 惰性 __getattr__ 兼容垫片,保留 from app.core.plugin import PluginManager(仓内 0 外部插件引用,纯防御性)。

- 16 个内部导入方改指 app.helper.plugin_manager(core/event.py 的 2 处保持函数级 lazy)。- 不依赖 chain/agent → 迁 helper 不造新环;垫片惰性 → 不引入顶层 core→helper 边。- 副作用:core 顶层 app.db 反向边清零(plugin.py 是最后 2 条所在)。

里程碑守护+收获:core 顶层 helper 边=0(解环里程碑保住)、core 顶层 db 边=0(core→db 直接边彻底清零)。4 个内省 seam 测试(S5a/S5b/S1b/S1c)retarget 至新文件;新增 relocation 测试;plugin/cycle/redis/metainfo 共 121 passed。

注:本 PR 仅 relocate(迁位置),god-object 职责拆分(Lifecycle/Market/Cloner/Metadata)留待后续 decompose。
