+++ app/agent/tools/factory.py
@@ -80,7 +80,7 @@ from app.agent.tools.impl.update_custom_identifiers import UpdateCustomIdentifie
 from app.agent.tools.impl.query_system_settings import QuerySystemSettingsTool
 from app.agent.tools.impl.update_system_settings import UpdateSystemSettingsTool
 from app.agent.llm.capability import AgentCapabilityManager
-from app.core.plugin import PluginManager
+from app.helper.plugin_manager import PluginManager
 from app.log import logger
 from app.schemas.message import ChannelCapabilityManager
 from app.schemas.types import MessageChannel
