+++ app/agent/tools/impl/query_plugin_capabilities.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel, Field
 
 from app.agent.tools.base import MoviePilotTool
 from app.agent.tools.tags import ToolTag
-from app.core.plugin import PluginManager
+from app.helper.plugin_manager import PluginManager
 from app.log import logger
