+++ app/chain/system.py
@@ -6,7 +6,7 @@ from typing import Union, Optional
 
 from app.chain import ChainBase
 from app.core.config import settings
-from app.core.plugin import PluginManager
+from app.helper.plugin_manager import PluginManager
 from app.helper.system import SystemHelper
 from app.log import logger
 from app.schemas import Notification, MessageChannel
