+++ app/workflow/actions/invoke_plugin.py
@@ -1,7 +1,7 @@
 from pydantic import Field
 
 from app.workflow.actions import BaseAction
-from app.core.plugin import PluginManager
+from app.helper.plugin_manager import PluginManager
 from app.log import logger
 from app.schemas import ActionParams, ActionContext
