+++ tests/test_core_auth_level.py
@@ -52,7 +52,7 @@ class CoreAuthLevelSeamTest(TestCase):
     def test_core_plugin_no_longer_imports_helper_sites(self):
         """app/core/plugin.py 的 auth_level 读取改走 seam,不再 import app.helper.sites
         （plugin.py 仍保留 helper.server / helper.plugin,属各自独立接缝）"""
-        import app.core.plugin as plugin
+        import app.helper.plugin_manager as plugin
         source = Path(plugin.__file__).read_text(encoding="utf-8")
         self.assertNotIn("from app.helper.sites", source)
         self.assertNotIn("import app.helper.sites", source)
